refactor(rest): rename internal-types to types
The types defined in "internal-types.ts" are not internal at all,
for example they include signatures of sequence actions.

This commit renames the file to "types.ts", which is a file name
already used in other packages (let's be consistent).
