+++ packages/rest/src/index.ts
@@ -35,7 +35,7 @@ export {writeResultToResponse} from './writer';
 export {HttpErrors};
 
 export * from './http-handler';
-export * from './internal-types';
+export * from './types';
 export * from './keys';
 export * from './rest.application';
 export * from './rest.component';
