+++ packages/rest/src/providers/reject.provider.ts
@@ -3,7 +3,7 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import {LogError, Reject} from '../internal-types';
+import {LogError, Reject} from '../types';
 import {inject, Provider} from '@loopback/context';
 import {ServerResponse, ServerRequest} from 'http';
 import {HttpError} from 'http-errors';
